# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: HelpMate API
  version: 0.1.0
paths:
...
  /api/proposals:
    get:
      tags:
      - proposals
      summary: List Proposals
      operationId: list_proposals_api_proposals_get
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/ProposalStatus'
          - type: 'null'
          title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Proposal'
                title: Response List Proposals Api Proposals Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/proposals/{proposal_id}:
    get:
      tags:
      - proposals
      summary: Get Proposal
      operationId: get_proposal_api_proposals__proposal_id__get
      parameters:
      - name: proposal_id
        in: path
        required: true
        schema:
          type: string
          title: Proposal Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Proposal'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/proposals/{proposal_id}/decision:
    post:
      tags:
      - proposals
      summary: Decide
      description: Approve, edit-then-approve, or reject. The only way an agent write ever executes.
      operationId: decide_api_proposals__proposal_id__decision_post
      parameters:
      - name: proposal_id
        in: path
        required: true
        schema:
          type: string
          title: Proposal Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Proposal'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/folders:
    get:
      tags:
      - organize
      summary: List Folders
      operationId: list_folders_api_folders_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Folder'
                type: array
                title: Response List Folders Api Folders Get
    post:
      tags:
      - organize
      summary: Create Folder
      operationId: create_folder_api_folders_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFolderIn'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Folder'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/folders/{folder_id}/items:
    get:
      tags:
      - organize
      summary: List Items
      operationId: list_items_api_folders__folder_id__items_get
      parameters:
      - name: folder_id
        in: path
        required: true
        schema:
          type: string
          title: Folder Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
                title: Response List Items Api Folders  Folder Id  Items Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - organize
      summary: Create Item
      operationId: create_item_api_folders__folder_id__items_post
      parameters:
      - name: folder_id
        in: path
        required: true
        schema:
          type: string
          title: Folder Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItemIn'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/tasks:
    get:
      tags:
      - organize
      summary: List Tasks
      operationId: list_tasks_api_tasks_get
      parameters:
      - name: horizon
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/Horizon'
          - type: 'null'
          title: Horizon
      - name: done
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Done
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
                title: Response List Tasks Api Tasks Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - organize
      summary: Create Task
      operationId: create_task_api_tasks_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskIn'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/tasks/{task_id}:
    patch:
      tags:
      - organize
      summary: Update Task
      operationId: update_task_api_tasks__task_id__patch
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          title: Task Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchTaskIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/reminders:
    get:
      tags:
      - organize
      summary: List Reminders
      operationId: list_reminders_api_reminders_get
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/ReminderStatus'
          - type: 'null'
          title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reminder'
                title: Response List Reminders Api Reminders Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/reminders/{reminder_id}/cancel:
    post:
      tags:
      - organize
      summary: Cancel Reminder
      operationId: cancel_reminder_api_reminders__reminder_id__cancel_post
      parameters:
      - name: reminder_id
        in: path
        required: true
        schema:
          type: string
          title: Reminder Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/push/subscriptions:
    post:
      tags:
      - push
      summary: Subscribe
      operationId: subscribe_api_push_subscriptions_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionIn'
        required: true
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - push
      summary: Unsubscribe
      operationId: unsubscribe_api_push_subscriptions_delete
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnsubscribeIn'
        required: true
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/push/ack:
    post:
      tags:
      - push
      summary: Ack
      description: 'Called by the service worker when a push arrives: this is how latency gets measured.'
      operationId: ack_api_push_ack_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AckIn'
        required: true
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AckIn:
      properties:
        notification_id:
          type: string
          title: Notification Id
        received_at:
          type: string
          format: date-time
          title: Received At
      additionalProperties: false
      type: object
      required:
      - notification_id
      - received_at
      title: AckIn
...
    CreateFolderIn:
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 1
          title: Name
        fields:
          items:
            $ref: '#/components/schemas/FieldDef'
          type: array
          title: Fields
          default: []
      additionalProperties: false
      type: object
      required:
      - name
      title: CreateFolderIn
    CreateItemIn:
      properties:
        title:
          type: string
          maxLength: 300
          minLength: 1
          title: Title
        fields:
          additionalProperties: true
          type: object
          title: Fields
          default: {}
      additionalProperties: false
      type: object
      required:
      - title
      title: CreateItemIn
...
    CreateTaskIn:
      properties:
        title:
          type: string
          maxLength: 300
          minLength: 1
          title: Title
        horizon:
          $ref: '#/components/schemas/Horizon'
          default: week
        folder_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Folder Id
        due_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Due At
      additionalProperties: false
      type: object
      required:
      - title
      title: CreateTaskIn
    DecisionIn:
      properties:
        decision:
          type: string
          enum:
          - approve
          - reject
          - edit
          title: Decision
        args:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Args
      additionalProperties: false
      type: object
      required:
      - decision
      title: DecisionIn
...
    FieldDef:
      properties:
        key:
          type: string
          pattern: ^[a-z][a-z0-9_]{0,39}$
          title: Key
        label:
          type: string
          title: Label
        type:
          $ref: '#/components/schemas/FieldType'
          default: text
        options:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Options
      additionalProperties: false
      type: object
      required:
      - key
      - label
      title: FieldDef
    FieldType:
      type: string
      enum:
      - text
      - number
      - date
      - select
      - rating
      - bool
      title: FieldType
    Folder:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        kind:
          $ref: '#/components/schemas/FolderKind'
          default: custom
        fields:
          items:
            $ref: '#/components/schemas/FieldDef'
          type: array
          title: Fields
          default: []
        created_at:
          type: string
          format: date-time
          title: Created At
      additionalProperties: false
      type: object
      required:
      - id
      - name
      - created_at
      title: Folder
    FolderKind:
      type: string
      enum:
      - para
      - custom
      title: FolderKind
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Horizon:
      type: string
      enum:
      - week
      - term
      - year
      - someday
      title: Horizon
    Item:
      properties:
        id:
          type: string
          title: Id
        folder_id:
          type: string
          title: Folder Id
        title:
          type: string
          title: Title
        fields:
          additionalProperties: true
          type: object
          title: Fields
          default: {}
        source:
          anyOf:
          - type: string
          - type: 'null'
          title: Source
        created_at:
          type: string
          format: date-time
          title: Created At
      additionalProperties: false
      type: object
      required:
      - id
      - folder_id
      - title
      - created_at
      title: Item
...
    PatchTaskIn:
      properties:
        title:
          anyOf:
          - type: string
            maxLength: 300
            minLength: 1
          - type: 'null'
          title: Title
        horizon:
          anyOf:
          - $ref: '#/components/schemas/Horizon'
          - type: 'null'
        done:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Done
        due_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Due At
      additionalProperties: false
      type: object
      title: PatchTaskIn
...
    Proposal:
      properties:
        id:
          type: string
          title: Id
        session_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
        tool:
          type: string
          title: Tool
        title:
          type: string
          title: Title
        summary:
          type: string
          title: Summary
        args:
          additionalProperties: true
          type: object
          title: Args
        preview:
          anyOf:
          - type: string
          - type: 'null'
          title: Preview
        risk:
          $ref: '#/components/schemas/Risk'
          default: write
        status:
          $ref: '#/components/schemas/ProposalStatus'
          default: pending
        created_at:
          type: string
          format: date-time
          title: Created At
        decided_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Decided At
        result:
          anyOf:
          - type: string
          - type: 'null'
          title: Result
      additionalProperties: false
      type: object
      required:
      - id
      - tool
      - title
      - summary
      - args
      - created_at
      title: Proposal
      description: A write the agent wants to make. Nothing runs until the owner approves it.
    ProposalStatus:
      type: string
      enum:
      - pending
      - approved
      - rejected
      - executed
      - failed
      title: ProposalStatus
    PushKeys:
      properties:
        p256dh:
          type: string
          title: P256Dh
        auth:
          type: string
          title: Auth
      additionalProperties: false
      type: object
      required:
      - p256dh
      - auth
      title: PushKeys
...
    Reminder:
      properties:
        id:
          type: string
          title: Id
        text:
          type: string
          title: Text
        due_at:
          type: string
          format: date-time
          title: Due At
        recurrence:
          anyOf:
          - type: string
          - type: 'null'
          title: Recurrence
        status:
          $ref: '#/components/schemas/ReminderStatus'
          default: scheduled
        created_at:
          type: string
          format: date-time
          title: Created At
        sent_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Sent At
      additionalProperties: false
      type: object
      required:
      - id
      - text
      - due_at
      - created_at
      title: Reminder
    ReminderStatus:
      type: string
      enum:
      - scheduled
      - sent
      - cancelled
      title: ReminderStatus
    Risk:
      type: string
      enum:
      - write
      - external
      title: Risk
...
    SubscriptionIn:
      properties:
        endpoint:
          type: string
          maxLength: 2000
          minLength: 10
          title: Endpoint
        keys:
          $ref: '#/components/schemas/PushKeys'
        user_agent:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: User Agent
      additionalProperties: false
      type: object
      required:
      - endpoint
      - keys
      title: SubscriptionIn
...
    Task:
      properties:
        id:
          type: string
          title: Id
        title:
          type: string
          title: Title
        horizon:
          $ref: '#/components/schemas/Horizon'
          default: week
        folder_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Folder Id
        due_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Due At
        done:
          type: boolean
          title: Done
          default: false
        created_at:
          type: string
          format: date-time
          title: Created At
      additionalProperties: false
      type: object
      required:
      - id
      - title
      - created_at
      title: Task
...
    UnsubscribeIn:
      properties:
        endpoint:
          type: string
          title: Endpoint
      additionalProperties: false
      type: object
      required:
      - endpoint
      title: UnsubscribeIn
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...